TASK1/requirements.txt
# Core dependencies
ultralytics>=8.0.200
opencv-python-headless>=4.8.1.78
numpy>=1.24.3
torch>=2.0.1
torchvision>=0.15.2

# Data processing and visualization
pandas>=2.0.3
matplotlib>=3.7.1
seaborn>=0.12.2
Pillow>=10.0.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
scikit-learn>=1.3.0
scipy>=1.11.1

# Additional tools
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# For video processing
opencv-contrib-python>=4.8.1.78

# For performance monitoring
psutil>=5.9.5
GPUtil>=1.4.0

# For CLI
argparse>=1.4.0
colorama>=0.4.6

# Optional: For web deployment
# flask>=2.3.2
# gunicorn>=21.2.0